Added ringbuffers, build against libvterm, etc.

+++ src/segments.c
@@ -14,7 +14,7 @@ void init_gdt()
    gdt_set_gate(3, 0, 0xFFFFFFFF, 0xFA, 0xCF); // User mode code segment
    gdt_set_gate(4, 0, 0xFFFFFFFF, 0xF2, 0xCF); // User mode data segment
 
-   gdt_flush(&gdt_ptr);
+   gdt_flush((uint32_t)&gdt_ptr);
 }
 void gdt_set_gate(int32_t num, uint32_t base, uint32_t limit, uint8_t access, uint8_t gran)
 {
